@@ -14,7 +14,7 @@ members = [
 
 [workspace.package]
 edition = "2024"
-rust-version = "1.86"
+rust-version = "1.87"
 license = "MIT"
 authors = ["Miden contributors"]
 homepage = "https://miden.xyz"
@@ -39,23 +39,19 @@ codegen-units = 1
 lto = true
 
 [workspace.dependencies]
-assembly = { package = "miden-assembly", version = "0.13", default-features = false }
+assembly = { package = "miden-assembly", version = "0.14", default-features = false }
 assert_matches = { version = "1.5", default-features = false }
-miden-block-prover = { path = "crates/miden-block-prover", version = "0.9", default-features = false }
+miden-block-prover = { path = "crates/miden-block-prover", version = "0.10", default-features = false }
 miden-crypto = { version = "0.14", default-features = false }
-miden-lib = { path = "crates/miden-lib", version = "0.9", default-features = false }
-miden-objects = { path = "crates/miden-objects", version = "0.9", default-features = false }
-miden-prover = { version = "0.13", default-features = false }
-miden-stdlib = { version = "0.13", default-features = false }
-miden-testing = { path = "crates/miden-testing", version = "0.9", default-features = false }
-miden-tx = { path = "crates/miden-tx", version = "0.9", default-features = false }
-miden-tx-batch-prover = { path = "crates/miden-tx-batch-prover", version = "0.9", default-features = false }
-miden-verifier = { version = "0.13", default-features = false }
+miden-lib = { path = "crates/miden-lib", version = "0.10", default-features = false }
+miden-objects = { path = "crates/miden-objects", version = "0.10", default-features = false }
+miden-prover = { version = "0.14", default-features = false }
+miden-stdlib = { version = "0.14", default-features = false }
+miden-testing = { path = "crates/miden-testing", version = "0.10", default-features = false }
+miden-tx = { path = "crates/miden-tx", version = "0.10", default-features = false }
+miden-tx-batch-prover = { path = "crates/miden-tx-batch-prover", version = "0.10", default-features = false }
+miden-verifier = { version = "0.14", default-features = false }
 rand = { version = "0.9", default-features = false }
 thiserror = { version = "2.0", default-features = false }
-vm-core = { package = "miden-core", version = "0.13", default-features = false }
-vm-processor = { package = "miden-processor", version = "0.13", default-features = false }
-
-# TODO: remove when pingora releases a fix https://github.com/cloudflare/pingora/issues/568
-[patch.crates-io]
-sfv = { git = "https://github.com/undef1nd/sfv.git", tag = "v0.9.4" }
+vm-core = { package = "miden-core", version = "0.14", default-features = false }
+vm-processor = { package = "miden-processor", version = "0.14", default-features = false }
