@@ -1,11 +1,7 @@
 use core::fmt;
-use std::{
-    fs::{File, read_to_string, write},
-    io::Write,
-    path::Path,
-    sync::Arc,
-};
+use std::{fs::File, io::Write, path::Path};
 
+use anyhow::Context;
 use miden_lib::{note::create_p2id_note, transaction::TransactionKernel};
 use miden_objects::{
     Felt,
@@ -13,10 +9,9 @@ use miden_objects::{
     asset::{Asset, FungibleAsset},
     crypto::rand::RpoRandomCoin,
     note::NoteType,
-    transaction::{TransactionArgs, TransactionMeasurements, TransactionScript},
+    transaction::{TransactionMeasurements, TransactionScript},
 };
 use miden_testing::TransactionContextBuilder;
-use miden_tx::TransactionExecutor;
 use vm_processor::ONE;
 
 mod utils;
@@ -39,11 +34,11 @@ impl fmt::Display for Benchmark {
     }
 }
 
-fn main() -> Result<(), String> {
+fn main() -> anyhow::Result<()> {
     // create a template file for benchmark results
     let path = Path::new("bin/bench-tx/bench-tx.json");
-    let mut file = File::create(path).map_err(|e| e.to_string())?;
-    file.write_all(b"{}").map_err(|e| e.to_string())?;
+    let mut file = File::create(path).context("failed to create file")?;
+    file.write_all(b"{}").context("failed to write to file")?;
 
     // run all available benchmarks
     let benchmark_results = vec![
@@ -61,27 +56,19 @@ fn main() -> Result<(), String> {
 // ================================================================================================
 
 /// Runs the default transaction with empty transaction script and two default notes.
-pub fn benchmark_default_tx() -> Result<TransactionMeasurements, String> {
+#[allow(clippy::arc_with_non_send_sync)]
+pub fn benchmark_default_tx() -> anyhow::Result<TransactionMeasurements> {
     let tx_context = TransactionContextBuilder::with_standard_account(ONE)
         .with_mock_notes_preserved()
         .build();
-
-    let account_id = tx_context.account().id();
-
-    let block_ref = tx_context.tx_inputs().block_header().block_num();
-    let tx_args = tx_context.tx_args().clone();
-    let notes = tx_context.tx_inputs().input_notes().clone();
-    let executor: TransactionExecutor =
-        TransactionExecutor::new(Arc::new(tx_context), None).with_tracing();
-    let executed_transaction = executor
-        .execute_transaction(account_id, block_ref, notes, tx_args)
-        .map_err(|e| e.to_string())?;
+    let executed_transaction = tx_context.execute().context("failed to execute transaction")?;
 
     Ok(executed_transaction.into())
 }
 
 /// Runs the transaction which consumes a P2ID note into a basic wallet.
-pub fn benchmark_p2id() -> Result<TransactionMeasurements, String> {
+#[allow(clippy::arc_with_non_send_sync)]
+pub fn benchmark_p2id() -> anyhow::Result<TransactionMeasurements> {
     // Create assets
     let faucet_id = AccountId::try_from(ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET).unwrap();
     let fungible_asset: Asset = FungibleAsset::new(faucet_id, 100).unwrap().into();
@@ -110,25 +97,17 @@ pub fn benchmark_p2id() -> Result<TransactionMeasurements, String> {
     )
     .unwrap();
 
-    let tx_context = TransactionContextBuilder::new(target_account.clone())
-        .input_notes(vec![note.clone()])
-        .build();
-    let block_ref = tx_context.tx_inputs().block_header().block_num();
-
-    let notes = tx_context.tx_inputs().input_notes().clone();
-
-    let executor =
-        TransactionExecutor::new(Arc::new(tx_context), Some(falcon_auth.clone())).with_tracing();
-
     let tx_script_target =
         TransactionScript::compile(DEFAULT_AUTH_SCRIPT, [], TransactionKernel::assembler())
             .unwrap();
-    let tx_args_target = TransactionArgs::default().with_tx_script(tx_script_target);
 
-    // execute transaction
-    let executed_transaction = executor
-        .execute_transaction(target_account.id(), block_ref, notes, tx_args_target)
-        .unwrap();
+    let tx_context = TransactionContextBuilder::new(target_account.clone())
+        .input_notes(vec![note])
+        .tx_script(tx_script_target)
+        .authenticator(Some(falcon_auth))
+        .build();
+
+    let executed_transaction = tx_context.execute()?;
 
     Ok(executed_transaction.into())
 }
