@@ -1,42 +1,28 @@
+use crate::errors::MasmError;
+
 // This file is generated by build.rs, do not modify manually.
 // It is generated by extracting errors from the masm files in the `miden-lib/asm` directory.
 //
 // To add a new error, define a constant in masm of the pattern `const.ERR_<CATEGORY>_...`.
 // Try to fit the error into a pre-existing category if possible (e.g. Account, Prologue,
 // Non-Fungible-Asset, ...).
-//
-// The comment directly above the constant will be interpreted as the error message for that error.
-
-// Note Script errors are in range 0x2c000..0x30000.
 
 // NOTE SCRIPT ERRORS
 // ================================================================================================
 
-/// P2ID script expects exactly 2 note inputs
-pub const ERR_P2ID_WRONG_NUMBER_OF_INPUTS: u32 = 0x2c000;
-/// P2ID's target account address and transaction address do not match
-pub const ERR_P2ID_TARGET_ACCT_MISMATCH: u32 = 0x2c001;
+/// Error Message: "P2IDR's reclaimer is not the original sender"
+pub const ERR_P2IDR_RECLAIM_ACCT_IS_NOT_SENDER: MasmError = MasmError::from_static_str("P2IDR's reclaimer is not the original sender");
+/// Error Message: "P2IDR can not be reclaimed as the transaction's reference block is lower than the reclaim height"
+pub const ERR_P2IDR_RECLAIM_HEIGHT_NOT_REACHED: MasmError = MasmError::from_static_str("P2IDR can not be reclaimed as the transaction's reference block is lower than the reclaim height");
+/// Error Message: "P2IDR scripts expect exactly 3 note inputs"
+pub const ERR_P2IDR_WRONG_NUMBER_OF_INPUTS: MasmError = MasmError::from_static_str("P2IDR scripts expect exactly 3 note inputs");
 
-/// P2IDR scripts expect exactly 3 note inputs
-pub const ERR_P2IDR_WRONG_NUMBER_OF_INPUTS: u32 = 0x2c002;
-/// P2IDR's reclaimer is not the original sender
-pub const ERR_P2IDR_RECLAIM_ACCT_IS_NOT_SENDER: u32 = 0x2c003;
-/// P2IDR can not be reclaimed as the transaction's reference block is lower than the reclaim height
-pub const ERR_P2IDR_RECLAIM_HEIGHT_NOT_REACHED: u32 = 0x2c004;
+/// Error Message: "P2ID's target account address and transaction address do not match"
+pub const ERR_P2ID_TARGET_ACCT_MISMATCH: MasmError = MasmError::from_static_str("P2ID's target account address and transaction address do not match");
+/// Error Message: "P2ID script expects exactly 2 note inputs"
+pub const ERR_P2ID_WRONG_NUMBER_OF_INPUTS: MasmError = MasmError::from_static_str("P2ID script expects exactly 2 note inputs");
 
-/// SWAP script expects exactly 10 note inputs
-pub const ERR_SWAP_WRONG_NUMBER_OF_INPUTS: u32 = 0x2c005;
-/// SWAP script requires exactly 1 note asset
-pub const ERR_SWAP_WRONG_NUMBER_OF_ASSETS: u32 = 0x2c006;
-
-pub const NOTE_SCRIPT_ERRORS: [(u32, &str); 7] = [
-    (ERR_P2ID_WRONG_NUMBER_OF_INPUTS, "P2ID script expects exactly 2 note inputs"),
-    (ERR_P2ID_TARGET_ACCT_MISMATCH, "P2ID's target account address and transaction address do not match"),
-
-    (ERR_P2IDR_WRONG_NUMBER_OF_INPUTS, "P2IDR scripts expect exactly 3 note inputs"),
-    (ERR_P2IDR_RECLAIM_ACCT_IS_NOT_SENDER, "P2IDR's reclaimer is not the original sender"),
-    (ERR_P2IDR_RECLAIM_HEIGHT_NOT_REACHED, "P2IDR can not be reclaimed as the transaction's reference block is lower than the reclaim height"),
-
-    (ERR_SWAP_WRONG_NUMBER_OF_INPUTS, "SWAP script expects exactly 10 note inputs"),
-    (ERR_SWAP_WRONG_NUMBER_OF_ASSETS, "SWAP script requires exactly 1 note asset"),
-];
+/// Error Message: "SWAP script requires exactly 1 note asset"
+pub const ERR_SWAP_WRONG_NUMBER_OF_ASSETS: MasmError = MasmError::from_static_str("SWAP script requires exactly 1 note asset");
+/// Error Message: "SWAP script expects exactly 10 note inputs"
+pub const ERR_SWAP_WRONG_NUMBER_OF_INPUTS: MasmError = MasmError::from_static_str("SWAP script expects exactly 10 note inputs");
