@@ -56,7 +56,7 @@ pub fn build_swap_tag(
 mod tests {
     use miden_objects::{
         self,
-        account::{AccountIdVersion, AccountStorageMode, AccountType, NetworkAccount},
+        account::{AccountIdVersion, AccountStorageMode, AccountType},
         asset::{FungibleAsset, NonFungibleAsset, NonFungibleAssetDetails},
     };
 
@@ -81,7 +81,6 @@ mod tests {
                     AccountIdVersion::Version0,
                     AccountType::FungibleFaucet,
                     AccountStorageMode::Public,
-                    NetworkAccount::Disabled,
                 ),
                 2500,
             )
@@ -96,7 +95,6 @@ mod tests {
                         AccountIdVersion::Version0,
                         AccountType::NonFungibleFaucet,
                         AccountStorageMode::Public,
-                        NetworkAccount::Disabled,
                     )
                     .prefix(),
                     vec![0xaa, 0xbb, 0xcc, 0xdd],
