@@ -1,13 +1,12 @@
 use alloc::{string::ToString, sync::Arc, vec::Vec};
 
 use miden_objects::{
-    Digest, EMPTY_WORD, Felt, TransactionInputError, TransactionOutputError, ZERO,
+    Digest, EMPTY_WORD, Felt, TransactionInputError, TransactionOutputError,
     account::{AccountCode, AccountId},
     assembly::{Assembler, DefaultSourceManager, KernelLibrary},
     block::BlockNumber,
     transaction::{
-        ForeignAccountInputs, OutputNote, OutputNotes, TransactionArgs, TransactionInputs,
-        TransactionOutputs,
+        OutputNote, OutputNotes, TransactionArgs, TransactionInputs, TransactionOutputs,
     },
     utils::{serde::Deserializable, sync::LazyLock},
     vm::{AdviceInputs, AdviceMap, Program, ProgramInfo, StackInputs, StackOutputs},
@@ -187,73 +186,6 @@ impl TransactionKernel {
         StackInputs::new(inputs)
             .map_err(|e| e.to_string())
             .expect("Invalid stack input")
-    }
-
-    /// Extends the advice inputs with the account-specific inputs.
-    pub fn extend_advice_inputs_for_account(
-        advice_inputs: &mut AdviceInputs,
-        foreign_account_inputs: &ForeignAccountInputs,
-    ) -> Result<(), TransactionInputError> {
-        let account_header = foreign_account_inputs.account_header();
-        let storage_header = foreign_account_inputs.storage_header();
-        let account_code = foreign_account_inputs.account_code();
-        let account_witness = foreign_account_inputs.witness();
-        let storage_proofs = foreign_account_inputs.storage_map_proofs();
-        let account_id = account_header.id();
-
-        // Note: keep in sync with the start_foreign_context kernel procedure
-        let account_key: Digest =
-            Digest::from([account_id.suffix(), account_id.prefix().as_felt(), ZERO, ZERO]);
-
-        // Extend the advice inputs with the new data
-        advice_inputs.extend_map([
-            // ACCOUNT_ID -> [ID_AND_NONCE, VAULT_ROOT, STORAGE_COMMITMENT, CODE_COMMITMENT]
-            (account_key, account_header.as_elements()),
-            // STORAGE_COMMITMENT -> [STORAGE_SLOT_DATA]
-            (account_header.storage_commitment(), storage_header.as_elements()),
-            // CODE_COMMITMENT -> [ACCOUNT_CODE_DATA]
-            (account_header.code_commitment(), account_code.as_elements()),
-        ]);
-
-        let account_leaf = account_witness.leaf();
-        let account_leaf_hash = account_leaf.hash();
-
-        // extend the merkle store and map with account witnesses merkle path
-        advice_inputs.extend_merkle_store(
-            account_witness
-                .path()
-                .inner_nodes(account_id.prefix().as_u64(), account_leaf_hash)
-                .map_err(|err| {
-                    TransactionInputError::InvalidMerklePath(
-                        format!("foreign account ID {}", account_id).into(),
-                        err,
-                    )
-                })?,
-        );
-
-        // populate advice map with the account's leaf
-        advice_inputs.extend_map([(account_leaf_hash, account_leaf.to_elements())]);
-
-        // Load merkle nodes for storage maps
-        for proof in storage_proofs {
-            // Extend the merkle store and map with the storage maps
-            advice_inputs.extend_merkle_store(
-                proof
-                    .path()
-                    .inner_nodes(proof.leaf().index().value(), proof.leaf().hash())
-                    .map_err(|err| {
-                        TransactionInputError::InvalidMerklePath(
-                            format!("foreign account ID {} storage proof", account_id).into(),
-                            err,
-                        )
-                    })?,
-            );
-            // Populate advice map with Sparse Merkle Tree leaf nodes
-            advice_inputs
-                .extend_map(core::iter::once((proof.leaf().hash(), proof.leaf().to_elements())));
-        }
-
-        Ok(())
     }
 
     /// Builds the stack for expected transaction execution outputs.
