@@ -5,7 +5,7 @@ use super::v0;
 use crate::{
     Felt,
     account::{
-        AccountIdV0, AccountIdVersion, AccountStorageMode, AccountType, NetworkAccount,
+        AccountIdV0, AccountIdVersion, AccountStorageMode, AccountType,
         account_id::AccountIdPrefixV0,
     },
     errors::AccountIdError,
@@ -117,22 +117,25 @@ impl AccountIdPrefix {
         }
     }
 
-    /// Returns true if an account with this ID is a public account.
+    /// Returns `true` if the full state of the account is on chain, i.e. if the modes are
+    /// [`AccountStorageMode::Public`] or [`AccountStorageMode::Network`], `false` otherwise.
+    pub fn is_onchain(&self) -> bool {
+        self.storage_mode().is_onchain()
+    }
+
+    /// Returns `true` if the storage mode is [`AccountStorageMode::Public`], `false` otherwise.
     pub fn is_public(&self) -> bool {
-        self.storage_mode() == AccountStorageMode::Public
+        self.storage_mode().is_public()
     }
 
-    /// Returns the network flag of this account, indicating whether self is a network account or
-    /// not.
-    pub fn network_account(&self) -> NetworkAccount {
-        match self {
-            AccountIdPrefix::V0(id_prefix) => id_prefix.network_account(),
-        }
-    }
-
-    /// Returns `true` if self is a network account, `false` otherwise.
+    /// Returns `true` if the storage mode is [`AccountStorageMode::Network`], `false` otherwise.
     pub fn is_network(&self) -> bool {
-        self.network_account().is_enabled()
+        self.storage_mode().is_network()
+    }
+
+    /// Returns `true` if self is a private account, `false` otherwise.
+    pub fn is_private(&self) -> bool {
+        self.storage_mode().is_private()
     }
 
     /// Returns the version of this account ID.
@@ -304,7 +307,7 @@ impl Deserializable for AccountIdPrefix {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::account::{AccountIdV0, NetworkAccount};
+    use crate::account::AccountIdV0;
 
     #[test]
     fn account_id_prefix_construction() {
@@ -320,25 +323,16 @@ mod tests {
                 AccountType::RegularAccountUpdatableCode,
             ] {
                 for storage_mode in [AccountStorageMode::Private, AccountStorageMode::Public] {
-                    for network_account in [NetworkAccount::Disabled, NetworkAccount::Enabled] {
-                        // Skip the invalid configuration.
-                        if !storage_mode.is_public() && network_account.is_enabled() {
-                            continue;
-                        }
+                    let id = AccountIdV0::dummy(input, account_type, storage_mode);
+                    let prefix = id.prefix();
+                    assert_eq!(prefix.account_type(), account_type);
+                    assert_eq!(prefix.storage_mode(), storage_mode);
+                    assert_eq!(prefix.version(), AccountIdVersion::Version0);
 
-                        let id =
-                            AccountIdV0::dummy(input, account_type, storage_mode, network_account);
-                        let prefix = id.prefix();
-                        assert_eq!(prefix.account_type(), account_type);
-                        assert_eq!(prefix.storage_mode(), storage_mode);
-                        assert_eq!(prefix.version(), AccountIdVersion::Version0);
-                        assert_eq!(prefix.network_account(), network_account);
-
-                        // Do a serialization roundtrip to ensure validity.
-                        let serialized_prefix = prefix.to_bytes();
-                        AccountIdPrefix::read_from_bytes(&serialized_prefix).unwrap();
-                        assert_eq!(serialized_prefix.len(), AccountIdPrefix::SERIALIZED_SIZE);
-                    }
+                    // Do a serialization roundtrip to ensure validity.
+                    let serialized_prefix = prefix.to_bytes();
+                    AccountIdPrefix::read_from_bytes(&serialized_prefix).unwrap();
+                    assert_eq!(serialized_prefix.len(), AccountIdPrefix::SERIALIZED_SIZE);
                 }
             }
         }
