@@ -1,6 +1,3 @@
-mod id_anchor;
-pub use id_anchor::AccountIdAnchor;
-
 pub(crate) mod v0;
 pub use v0::{AccountIdPrefixV0, AccountIdV0};
 
@@ -20,9 +17,6 @@ pub use account_type::AccountType;
 
 mod storage_mode;
 pub use storage_mode::AccountStorageMode;
-
-mod network_account;
-pub use network_account::NetworkAccount;
 
 mod id_version;
 use alloc::string::{String, ToString};
@@ -48,26 +42,25 @@ use crate::{AccountError, errors::AccountIdError};
 /// second is called the suffix. It is layed out as follows:
 ///
 /// ```text
-/// prefix: [hash (29 bits) | network account (1 bit) | hash (26 bits) |
-///          storage mode (2 bits) | type (2 bits) | version (4 bits)]
-/// suffix: [anchor_epoch (16 bits) | hash (40 bits) | 8 zero bits]
+/// prefix: [hash (56 bits) | storage mode (2 bits) | type (2 bits) | version (4 bits)]
+/// suffix: [zero bit | hash (55 bits) | 8 zero bits]
 /// ```
 ///
 /// # Generation
 ///
-/// An `AccountId` is a commitment to a user-generated seed, the code and storage of an account and
-/// to a certain hash of an epoch block of the blockchain. An id is generated by picking an epoch
-/// block as an anchor - which is why it is also referred to as the anchor block - and creating the
-/// account's initial storage and code. Then a random seed is picked and the hash of `(SEED,
-/// CODE_COMMITMENT, STORAGE_COMMITMENT, ANCHOR_BLOCK_COMMITMENT)` is computed. If the hash's first
-/// element has the desired storage mode, account type and version, the computation part of the ID
-/// generation is done. If not, another random seed is picked and the process is repeated. The first
-/// felt of the ID is then the first element of the hash.
+/// An `AccountId` is a commitment to a user-generated seed and the code and storage of an account.
+/// An id is generated by first creating the account's initial storage and code. Then a random seed
+/// is picked and the hash of `(SEED, CODE_COMMITMENT, STORAGE_COMMITMENT, EMPTY_WORD)` is computed.
+/// This process is repeated until the hash's first element has the desired storage mode, account
+/// type and version and the suffix' most significant bit is zero.
 ///
-/// The suffix of the ID is the second element of the hash. Its upper 16 bits are overwritten
-/// with the epoch in which the ID is anchored and the lower 8 bits are zeroed. Thus, the prefix
-/// of the ID must derive exactly from the hash, while only part of the suffix is derived from
-/// the hash.
+/// The prefix of the ID is exactly the first element of the hash. The suffix of the ID is the
+/// second element of the hash, but its lower 8 bits are zeroed. Thus, the prefix of the ID must
+/// derive exactly from the hash, while only the first 56 bits of the suffix are derived from the
+/// hash.
+///
+/// In total, due to requiring specific bits for storage mode, type, version and the most
+/// significant bit in the suffix, generating an ID requires 9 bits of Proof-of-Work.
 ///
 /// # Constraints
 ///
@@ -75,7 +68,7 @@ use crate::{AccountError, errors::AccountIdError};
 ///
 /// - The prefix contains account ID metadata (storage mode, type or version) that does not match
 ///   any of the known values.
-/// - The anchor epoch in the suffix is equal to [`u16::MAX`].
+/// - The most significant bit of the suffix is not zero.
 /// - The lower 8 bits of the suffix are not zero, although [`AccountId::new`] ensures this is the
 ///   case rather than return an error.
 ///
@@ -85,49 +78,20 @@ use crate::{AccountError, errors::AccountIdError};
 ///
 /// - The prefix is the output of a hash function so it will be a valid field element without
 ///   requiring additional constraints.
-/// - The network flag is placed at the 30th most significant bit in the prefix. This means that the
-///   last bit of a note tag with the `0b00` prefix will indicate whether the target account is a
-///   network account or not. Note tags intended for local execution use only the 14 most
-///   significant bits, so the network bit does not appear in those. This is desirable because if it
-///   was there, it would reduce the uniqueness of the note tag and it would be less useful for
-///   scanning for available notes in the network for private accounts.
 /// - The version is placed at a static offset such that future ID versions which may change the
 ///   number of type or storage mode bits will not cause the version to be at a different offset.
 ///   This is important so that a parser can always reliably read the version and then parse the
 ///   remainder of the ID depending on the version. Having only 4 bits for the version is a trade
-///   off between future proofing to be able to introduce more versions and the version requiring
-///   Proof of Work as part of the ID generation.
+///   off between future proofing to allow introducing more versions and the version requiring Proof
+///   of Work as part of the ID generation.
 /// - The version, type and storage mode are part of the prefix which is included in the
 ///   representation of a non-fungible asset. The prefix alone is enough to determine all of these
 ///   properties about the ID.
-///     - The anchor epoch is not important beyond the creation process, so placing it in the second
-///       felt is fine. Moreover, all properties of the prefix must be derived from the seed, so
-///       they add to the proof of work difficulty. Adding 16 bits of PoW for the epoch would be
-///       significant.
-/// - The anchor epoch is placed at the most significant end of the suffix. Its value must be less
-///   than [`u16::MAX`] so that at least one of the upper 16 bits is always zero. This ensures that
-///   the entire suffix is valid even if the remaining bits of the felt are one.
-/// - The lower 8 bits of the suffix may be overwritten when the ID is encoded in other layouts such
-///   as the [`NoteMetadata`](crate::note::NoteMetadata). In such cases, it can happen that all bits
-///   of the encoded suffix would be one, so having the epoch constraint is important.
-/// - The ID is dependent on the hash of an epoch block. This is a block whose number is a multiple
-///   of 2^[`BlockNumber::EPOCH_LENGTH_EXPONENT`][epoch_len_exp], e.g. `0`, `65536`, `131072`, ...
-///   These are the first blocks of epoch 0, 1, 2, ... We call this dependence _anchoring_ because
-///   the ID is anchored to that epoch block's commitment. Anchoring makes it practically impossible
-///   for an attacker to construct a rainbow table of account IDs whose epoch is X, if the block for
-///   epoch X has not been constructed yet because its hash is then unknown. Therefore, picking a
-///   recent anchor block when generating a new ID makes it extremely unlikely that an attacker can
-///   highjack this ID because the hash of that block has only been known for a short period of
-///   time.
-///     - An ID highjack refers to an attack where a user generates an ID and lets someone else send
-///       assets to it. At this point the user has not registered the ID on-chain yet, likely
-///       because they need the funds in the asset to pay for their first transaction where the
-///       account would be registered. Until the ID is registered on chain, an attacker with a
-///       rainbow table who happens to have a seed, code and storage commitment combination that
-///       hashes to the user's ID can claim the assets sent to the user's ID. Adding the anchor
-///       block commitment to the ID generation process makes this attack practically impossible.
-///
-/// [epoch_len_exp]: crate::block::BlockNumber::EPOCH_LENGTH_EXPONENT
+/// - The most significant bit of the suffix must be zero to ensure the value of the suffix is
+///   always a valid felt, even if the lower 8 bits are all set to `1`. The lower 8 bits of the
+///   suffix may be overwritten when the ID is embedded in other layouts such as the
+///   [`NoteMetadata`](crate::note::NoteMetadata). In that case, it can happen that all lower bits
+///   of the encoded suffix are one, so having the zero bit constraint is important for validity.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum AccountId {
     V0(AccountIdV0),
@@ -144,14 +108,8 @@ impl AccountId {
     // --------------------------------------------------------------------------------------------
 
     /// Creates an [`AccountId`] by hashing the given `seed`, `code_commitment`,
-    /// `storage_commitment` and [`AccountIdAnchor::block_commitment`] from the `anchor` and using
-    /// the resulting first and second element of the hash as the prefix and suffix felts of the
-    /// ID.
-    ///
-    /// The [`AccountIdAnchor::epoch`] from the `anchor` overwrites part of the suffix.
-    ///
-    /// Note that the `anchor` must correspond to a valid block in the chain for the ID to be deemed
-    /// valid during creation.
+    /// `storage_commitment` and using the resulting first and second element of the hash as the
+    /// prefix and suffix felts of the ID.
     ///
     /// See the documentation of the [`AccountId`] for more details on the generation.
     ///
@@ -161,20 +119,19 @@ impl AccountId {
     /// documentation](AccountId#constraints) for details.
     pub fn new(
         seed: Word,
-        anchor: AccountIdAnchor,
         version: AccountIdVersion,
         code_commitment: Digest,
         storage_commitment: Digest,
     ) -> Result<Self, AccountIdError> {
         match version {
             AccountIdVersion::Version0 => {
-                AccountIdV0::new(seed, anchor, code_commitment, storage_commitment).map(Self::V0)
+                AccountIdV0::new(seed, code_commitment, storage_commitment).map(Self::V0)
             },
         }
     }
 
     /// Creates an [`AccountId`] from the given felts where the felt at index 0 is the prefix
-    /// and the felt at index 2 is the suffix.
+    /// and the felt at index 1 is the suffix.
     ///
     /// # Warning
     ///
@@ -198,67 +155,55 @@ impl AccountId {
     }
 
     /// Constructs an [`AccountId`] for testing purposes with the given account type, storage
-    /// mode and network flag.
+    /// mode.
     ///
     /// This function does the following:
     /// - Split the given bytes into a `prefix = bytes[0..8]` and `suffix = bytes[8..]` part to be
     ///   used for the prefix and suffix felts, respectively.
     /// - The least significant byte of the prefix is set to the given version, type and storage
     ///   mode.
-    /// - The 30th most significant bit in the prefix is set to the network flag.
     /// - The 32nd most significant bit in the prefix is cleared to ensure it is a valid felt. The
     ///   32nd is chosen as it is the lowest bit that we can clear and still ensure felt validity.
-    ///   This leaves the upper 31 bits (except for the network flag) to be set by the input `bytes`
-    ///   which makes it simpler to create test values which more often need specific values for the
-    ///   most significant end of the ID.
-    /// - In the suffix the anchor epoch is set to 0 and the lower 8 bits are cleared.
-    ///
-    /// # Panics
-    ///
-    /// Panics if the network flag is set to [`NetworkAccount::Enabled`] and the storage mode is
-    /// [`AccountStorageMode::Private`].
+    ///   This leaves the upper 31 bits to be set by the input `bytes` which makes it simpler to
+    ///   create test values which more often need specific values for the most significant end of
+    ///   the ID.
+    /// - In the suffix the most significant bit and the lower 8 bits are cleared.
     #[cfg(any(feature = "testing", test))]
     pub fn dummy(
         bytes: [u8; 15],
         version: AccountIdVersion,
         account_type: AccountType,
         storage_mode: AccountStorageMode,
-        network_account: NetworkAccount,
     ) -> AccountId {
         match version {
             AccountIdVersion::Version0 => {
-                Self::V0(AccountIdV0::dummy(bytes, account_type, storage_mode, network_account))
+                Self::V0(AccountIdV0::dummy(bytes, account_type, storage_mode))
             },
         }
     }
 
     /// Grinds an account seed until its hash matches the given `account_type`, `storage_mode` and
     /// `version` and returns it as a [`Word`]. The input to the hash function next to the seed are
-    /// the `code_commitment`, `storage_commitment` and `anchor_block_commitment`.
+    /// the `code_commitment` and `storage_commitment`.
     ///
     /// The grinding process is started from the given `init_seed` which should be a random seed
     /// generated from a cryptographically secure source.
-    #[allow(clippy::too_many_arguments)]
     pub fn compute_account_seed(
         init_seed: [u8; 32],
         account_type: AccountType,
         storage_mode: AccountStorageMode,
-        network_account: NetworkAccount,
         version: AccountIdVersion,
         code_commitment: Digest,
         storage_commitment: Digest,
-        anchor_block_commitment: Digest,
     ) -> Result<Word, AccountError> {
         match version {
             AccountIdVersion::Version0 => AccountIdV0::compute_account_seed(
                 init_seed,
                 account_type,
                 storage_mode,
-                network_account,
                 version,
                 code_commitment,
                 storage_commitment,
-                anchor_block_commitment,
             ),
         }
     }
@@ -290,9 +235,20 @@ impl AccountId {
         }
     }
 
+    /// Returns `true` if the full state of the account is on chain, i.e. if the modes are
+    /// [`AccountStorageMode::Public`] or [`AccountStorageMode::Network`], `false` otherwise.
+    pub fn is_onchain(&self) -> bool {
+        self.storage_mode().is_onchain()
+    }
+
     /// Returns `true` if the storage mode is [`AccountStorageMode::Public`], `false` otherwise.
     pub fn is_public(&self) -> bool {
         self.storage_mode().is_public()
+    }
+
+    /// Returns `true` if the storage mode is [`AccountStorageMode::Network`], `false` otherwise.
+    pub fn is_network(&self) -> bool {
+        self.storage_mode().is_network()
     }
 
     /// Returns `true` if the storage mode is [`AccountStorageMode::Private`], `false` otherwise.
@@ -300,33 +256,10 @@ impl AccountId {
         self.storage_mode().is_private()
     }
 
-    /// Returns the network flag of this account, indicating whether self is a network account or
-    /// not.
-    pub fn network_account(&self) -> NetworkAccount {
-        match self {
-            AccountId::V0(account_id) => account_id.network_account(),
-        }
-    }
-
-    /// Returns `true` if self is a network account, `false` otherwise.
-    pub fn is_network(&self) -> bool {
-        self.network_account().is_enabled()
-    }
-
     /// Returns the version of this account ID.
     pub fn version(&self) -> AccountIdVersion {
         match self {
             AccountId::V0(_) => AccountIdVersion::Version0,
-        }
-    }
-
-    /// Returns the anchor epoch of this account ID.
-    ///
-    /// This is the epoch to which this ID is anchored. The hash of this epoch block is used in the
-    /// generation of the ID.
-    pub fn anchor_epoch(&self) -> u16 {
-        match self {
-            AccountId::V0(account_id) => account_id.anchor_epoch(),
         }
     }
 
