@@ -27,7 +27,7 @@ pub use errors::{
     AccountDeltaError, AccountError, AccountIdError, AccountTreeError, AssetError, AssetVaultError,
     BatchAccountUpdateError, NetworkIdError, NoteError, NullifierTreeError, PartialBlockchainError,
     ProposedBatchError, ProposedBlockError, ProvenBatchError, ProvenTransactionError,
-    TransactionInputError, TransactionOutputError, TransactionScriptError,
+    TokenSymbolError, TransactionInputError, TransactionOutputError, TransactionScriptError,
 };
 pub use miden_crypto::hash::rpo::{Rpo256 as Hasher, RpoDigest as Digest};
 pub use vm_core::{
@@ -41,7 +41,7 @@ pub mod assembly {
         Assembler, AssemblyError, Compile, CompileOptions, DefaultSourceManager, KernelLibrary,
         Library, LibraryNamespace, LibraryPath, SourceManager, Version,
         ast::{Module, ModuleKind, ProcedureName, QualifiedProcedureName},
-        mast,
+        diagnostics, mast,
     };
 }
 
