@@ -1,8 +1,6 @@
 use rand_xoshiro::rand_core::SeedableRng;
 
-use crate::account::{
-    AccountId, AccountIdV0, AccountIdVersion, AccountStorageMode, AccountType, NetworkAccount,
-};
+use crate::account::{AccountId, AccountIdV0, AccountIdVersion, AccountStorageMode, AccountType};
 
 // CONSTANTS
 // --------------------------------------------------------------------------------------------
@@ -47,9 +45,9 @@ pub const ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_UPDATABLE_CODE_ON_CHAIN_2: u128 = ac
 // REGULAR ACCOUNTS - NETWORK
 pub const ACCOUNT_ID_REGULAR_NETWORK_ACCOUNT_IMMUTABLE_CODE: u128 = account_id(
     AccountType::RegularAccountImmutableCode,
-    AccountStorageMode::Public,
+    AccountStorageMode::Network,
     0xaacc_bbdd,
-) | network_account_bitmask();
+);
 
 // These faucet IDs all have a unique prefix and suffix felts. This is to ensure that when they
 // are used to issue an asset they don't cause us to run into the "multiple leaf" case when
@@ -70,8 +68,7 @@ pub const ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET_3: u128 =
     account_id(AccountType::FungibleFaucet, AccountStorageMode::Public, 0xeeff_cc99);
 // FUNGIBLE TOKENS - NETWORK
 pub const ACCOUNT_ID_NETWORK_FUNGIBLE_FAUCET: u128 =
-    account_id(AccountType::FungibleFaucet, AccountStorageMode::Public, 0xaabc_bcdf)
-        | network_account_bitmask();
+    account_id(AccountType::FungibleFaucet, AccountStorageMode::Network, 0xaabc_bcdf);
 
 // NON-FUNGIBLE TOKENS - PRIVATE
 pub const ACCOUNT_ID_PRIVATE_NON_FUNGIBLE_FAUCET: u128 =
@@ -83,8 +80,7 @@ pub const ACCOUNT_ID_PUBLIC_NON_FUNGIBLE_FAUCET_1: u128 =
     account_id(AccountType::NonFungibleFaucet, AccountStorageMode::Public, 0xccdf_eefa);
 // NON-FUNGIBLE TOKENS - NETWORK
 pub const ACCOUNT_ID_NETWORK_NON_FUNGIBLE_FAUCET: u128 =
-    account_id(AccountType::NonFungibleFaucet, AccountStorageMode::Public, 0xabbc_ffde)
-        | network_account_bitmask();
+    account_id(AccountType::NonFungibleFaucet, AccountStorageMode::Network, 0xabbc_ffde);
 
 // TEST ACCOUNT IDs WITH CERTAIN PROPERTIES
 /// The Account Id with the maximum possible one bits.
@@ -101,8 +97,6 @@ pub const ACCOUNT_ID_MAX_ZEROES: u128 =
 /// Produces a valid account ID with the given account type and storage mode.
 ///
 /// - Version is set to 0.
-/// - Network flag is set to 0, i.e. disabled (the account is not a network account).
-/// - Anchor epoch is set to 0.
 ///
 /// Finally, distributes the given `random` value over the ID to produce non-trivial values for
 /// testing. This is easiest explained with an example. Suppose `random` is `0xaabb_ccdd`,
@@ -110,7 +104,7 @@ pub const ACCOUNT_ID_MAX_ZEROES: u128 =
 ///
 /// ```text
 /// prefix: [0xaa | 5 zero bytes | 0xbb | metadata byte]
-/// suffix: [2 zero bytes (epoch) | 0xcc | 3 zero bytes | 0xdd | zero byte]
+/// suffix: [2 zero bytes | 0xcc | 3 zero bytes | 0xdd | zero byte]
 /// ```
 pub const fn account_id(
     account_type: AccountType,
@@ -140,12 +134,6 @@ pub const fn account_id(
     id
 }
 
-/// Returns the bitmask for the network flag. A bitwise OR operation on an account ID an this
-/// bitmask will set it.
-const fn network_account_bitmask() -> u128 {
-    (AccountIdV0::NETWORK_ACCOUNT_MASK as u128) << 64
-}
-
 /// A builder for creating [`AccountId`]s for testing purposes.
 ///
 /// This is essentially a wrapper around [`AccountId::dummy`] generating random values as its input.
@@ -154,7 +142,7 @@ const fn network_account_bitmask() -> u128 {
 /// # Example
 ///
 /// ```
-/// # use miden_objects::account::{AccountType, AccountStorageMode, AccountId, NetworkAccount};
+/// # use miden_objects::account::{AccountType, AccountStorageMode, AccountId};
 /// # use miden_objects::testing::account_id::{AccountIdBuilder};
 ///
 /// let mut rng = rand::rng();
@@ -169,22 +157,16 @@ const fn network_account_bitmask() -> u128 {
 ///     .build_with_rng(&mut rng);
 /// assert_eq!(random_id2.account_type(), AccountType::FungibleFaucet);
 /// assert_eq!(random_id2.storage_mode(), AccountStorageMode::Private);
-/// assert_eq!(random_id2.network_account(), NetworkAccount::Disabled);
 /// ```
 pub struct AccountIdBuilder {
     account_type: Option<AccountType>,
     storage_mode: Option<AccountStorageMode>,
-    network_account: Option<NetworkAccount>,
 }
 
 impl AccountIdBuilder {
     /// Creates a new [`AccountIdBuilder`].
     pub fn new() -> Self {
-        Self {
-            account_type: None,
-            storage_mode: None,
-            network_account: None,
-        }
+        Self { account_type: None, storage_mode: None }
     }
 
     /// Sets the [`AccountType`] of the generated [`AccountId`] to the provided value.
@@ -199,22 +181,10 @@ impl AccountIdBuilder {
         self
     }
 
-    /// Sets the [`NetworkAccount`] of the generated [`AccountId`] to the provided value.
-    pub fn network_account(mut self, network_account: NetworkAccount) -> Self {
-        self.network_account = Some(network_account);
-        self
-    }
-
     /// Builds an [`AccountId`] using the provided [`rand::Rng`].
     ///
     /// If no [`AccountType`] or [`AccountStorageMode`] were previously set, random ones are
     /// generated.
-    ///
-    /// # Panics
-    ///
-    /// Panics if the network flag is set to [`NetworkAccount::Enabled`] and the storage mode is
-    /// [`AccountStorageMode::Private`]. This can only happen when the network flag is set
-    /// explicitly.
     pub fn build_with_rng<R: rand::Rng + ?Sized>(self, rng: &mut R) -> AccountId {
         let account_type = match self.account_type {
             Some(account_type) => account_type,
@@ -226,23 +196,7 @@ impl AccountIdBuilder {
             None => rng.random(),
         };
 
-        let network_account = match (self.network_account, storage_mode) {
-            // Do not automatically rewrite an incorrect configuration (enabled, private) to a
-            // correct one, so the caller is informed about the misconfiguration rather than
-            // it silently doing something unexpected.
-            (Some(network_account), _) => network_account,
-            (None, AccountStorageMode::Public) => rng.random(),
-            (None, AccountStorageMode::Private) => NetworkAccount::Disabled,
-        };
-
-        // This will panic if the network flag and storage mode are misconfigured.
-        AccountId::dummy(
-            rng.random(),
-            AccountIdVersion::Version0,
-            account_type,
-            storage_mode,
-            network_account,
-        )
+        AccountId::dummy(rng.random(), AccountIdVersion::Version0, account_type, storage_mode)
     }
 
     /// Builds an [`AccountId`] using the provided seed as input for an RNG implemented in
