@@ -18,9 +18,7 @@ mod tx_header;
 mod tx_witness;
 
 pub use executed_tx::{ExecutedTransaction, TransactionMeasurements};
-pub use inputs::{
-    ForeignAccountInputs, InputNote, InputNotes, ToInputNoteCommitments, TransactionInputs,
-};
+pub use inputs::{AccountInputs, InputNote, InputNotes, ToInputNoteCommitments, TransactionInputs};
 pub use ordered_transactions::OrderedTransactionHeaders;
 pub use outputs::{OutputNote, OutputNotes, TransactionOutputs};
 pub use partial_blockchain::PartialBlockchain;
