@@ -28,13 +28,17 @@ mod test_tx;
 
 #[macro_export]
 macro_rules! assert_execution_error {
-    ($execution_result:expr, $expected_err_code:expr) => {
+    ($execution_result:expr, $expected_err:expr) => {
         match $execution_result {
-            Err(vm_processor::ExecutionError::FailedAssertion { clk: _, err_code, err_msg: _ }) => {
-                assert!(
-                    err_code == $expected_err_code,
-                    "Execution failed on assertion with an unexpected error code (Actual err_code: 0x{:x}, expected 0x{:x}).",
-                    err_code, $expected_err_code
+            Err(vm_processor::ExecutionError::FailedAssertion { label: _, source_file: _, clk: _, err_code, err_msg }) => {
+                if let Some(ref msg) = err_msg {
+                  assert_eq!(msg.as_ref(), $expected_err.message(), "error messages did not match");
+                }
+
+                assert_eq!(
+                    err_code, $expected_err.code(),
+                    "Execution failed on assertion with an unexpected error (Actual code: {}, msg: {}, Expected code: {}).",
+                    err_code, err_msg.as_ref().map(|string| string.as_ref()).unwrap_or("<no message>"), $expected_err,
                 );
             },
             Ok(_) => panic!("Execution was unexpectedly successful"),
