@@ -4,7 +4,7 @@ use anyhow::Context;
 use miden_lib::{
     account::wallets::BasicWallet,
     errors::tx_kernel_errors::{
-        ERR_ACCOUNT_SEED_ANCHOR_BLOCK_COMMITMENT_DIGEST_MISMATCH,
+        ERR_ACCOUNT_SEED_AND_COMMITMENT_DIGEST_MISMATCH,
         ERR_PROLOGUE_NEW_FUNGIBLE_FAUCET_RESERVED_SLOT_MUST_BE_EMPTY,
         ERR_PROLOGUE_NEW_NON_FUNGIBLE_FAUCET_RESERVED_SLOT_MUST_BE_VALID_EMPY_SMT,
     },
@@ -30,16 +30,15 @@ use miden_lib::{
 };
 use miden_objects::{
     account::{
-        Account, AccountBuilder, AccountId, AccountIdAnchor, AccountIdVersion,
-        AccountProcedureInfo, AccountStorageMode, AccountType, NetworkAccount, StorageSlot,
+        Account, AccountBuilder, AccountId, AccountIdVersion, AccountProcedureInfo,
+        AccountStorageMode, AccountType, StorageSlot,
     },
-    block::{BlockHeader, BlockNumber},
     testing::{
         account_component::AccountMockComponent,
         account_id::{ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET, ACCOUNT_ID_PUBLIC_NON_FUNGIBLE_FAUCET},
         constants::FUNGIBLE_FAUCET_INITIAL_BALANCE,
     },
-    transaction::{ForeignAccountInputs, TransactionArgs, TransactionScript},
+    transaction::{AccountInputs, TransactionArgs, TransactionScript},
 };
 use rand::{Rng, SeedableRng};
 use rand_chacha::ChaCha20Rng;
@@ -92,7 +91,7 @@ fn test_transaction_prologue() {
         Some(tx_script),
         Some(note_args_map),
         tx_context.tx_args().advice_inputs().clone().map,
-        Vec::<ForeignAccountInputs>::new(),
+        Vec::<AccountInputs>::new(),
     );
 
     tx_context.set_tx_args(tx_args);
@@ -448,7 +447,6 @@ pub fn create_account_test(
 
 pub fn create_multiple_accounts_test(
     mock_chain: &MockChain,
-    anchor_block_header: &BlockHeader,
     storage_mode: AccountStorageMode,
 ) -> anyhow::Result<()> {
     let mut accounts = Vec::new();
@@ -462,10 +460,6 @@ pub fn create_multiple_accounts_test(
         let (account, seed) = AccountBuilder::new(ChaCha20Rng::from_os_rng().random())
             .account_type(account_type)
             .storage_mode(storage_mode)
-            .anchor(
-                AccountIdAnchor::try_from(anchor_block_header)
-                    .context("block header to anchor conversion failed")?,
-            )
             .with_component(
                 AccountMockComponent::new_with_slots(
                     TransactionKernel::testing_assembler(),
@@ -490,71 +484,34 @@ pub fn create_multiple_accounts_test(
     Ok(())
 }
 
-/// Tests that a valid account of each type can be created successfully with the genesis block used
-/// as the anchor block for the account IDs.
+/// Tests that a valid account of each storage mode can be created successfully.
 #[test]
-pub fn create_accounts_with_anchor_block_zero() -> anyhow::Result<()> {
-    let mut mock_chain = MockChain::new();
-    // Choose epoch block 0 as the anchor block.
-    // Here the transaction reference block is also the anchor block.
-    let genesis_block_header = mock_chain.block_header(BlockNumber::GENESIS.as_usize());
+pub fn create_accounts_with_all_storage_modes() -> anyhow::Result<()> {
+    let mock_chain = MockChain::new();
 
-    create_multiple_accounts_test(&mock_chain, &genesis_block_header, AccountStorageMode::Private)?;
+    create_multiple_accounts_test(&mock_chain, AccountStorageMode::Private)?;
 
-    // Seal one more block to test the case where the transaction reference block is not the anchor
-    // block.
-    mock_chain.prove_next_block();
+    create_multiple_accounts_test(&mock_chain, AccountStorageMode::Public)?;
 
-    create_multiple_accounts_test(&mock_chain, &genesis_block_header, AccountStorageMode::Public)
+    create_multiple_accounts_test(&mock_chain, AccountStorageMode::Network)
 }
 
-/// Tests that a valid account of each type can be created successfully with an epoch block whose
-/// number is non-zero used as the anchor block for the account IDs.
-///
-/// Note that this test is very slow in debug mode.
-#[test]
-pub fn create_accounts_with_non_zero_anchor_block() -> anyhow::Result<()> {
-    let mut mock_chain = MockChain::new();
-    mock_chain
-        .prove_until_block(1u32 << 16)
-        .context("failed to prove multiple blocks")?;
-
-    // Choose epoch block 1 whose block number is 2^16 as the anchor block.
-    // Here the transaction reference block is also the anchor block.
-    let epoch1_block_header = mock_chain.block_header(1 << 16);
-
-    create_multiple_accounts_test(&mock_chain, &epoch1_block_header, AccountStorageMode::Private)?;
-
-    // Seal one more block to test the case where the transaction reference block is not the anchor
-    // block.
-    mock_chain.prove_next_block();
-
-    create_multiple_accounts_test(&mock_chain, &epoch1_block_header, AccountStorageMode::Public)
-}
-
-/// Takes an account with a placeholder ID and returns the same account but with its ID replaced.
-/// The ID is newly generated and anchored in the given block header.
-fn compute_valid_account_id(
-    account: Account,
-    anchor_block_header: &BlockHeader,
-) -> (Account, Word) {
+/// Takes an account with a placeholder ID and returns the same account but with its ID replaced
+/// with a newly generated one.
+fn compute_valid_account_id(account: Account) -> (Account, Word) {
     let init_seed: [u8; 32] = [5; 32];
     let seed = AccountId::compute_account_seed(
         init_seed,
         account.account_type(),
         AccountStorageMode::Public,
-        NetworkAccount::Disabled,
         AccountIdVersion::Version0,
         account.code().commitment(),
         account.storage().commitment(),
-        anchor_block_header.commitment(),
     )
     .unwrap();
 
-    let anchor = AccountIdAnchor::try_from(anchor_block_header).unwrap();
     let account_id = AccountId::new(
         seed,
-        anchor,
         AccountIdVersion::Version0,
         account.code().commitment(),
         account.storage().commitment(),
@@ -575,15 +532,13 @@ pub fn create_account_fungible_faucet_invalid_initial_balance() -> anyhow::Resul
     let mut mock_chain = MockChain::new();
     mock_chain.prove_next_block();
 
-    let genesis_block_header = mock_chain.block_header(BlockNumber::GENESIS.as_usize());
-
     let account = Account::mock_fungible_faucet(
         ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET,
         ZERO,
         Felt::new(FUNGIBLE_FAUCET_INITIAL_BALANCE),
         TransactionKernel::assembler().with_debug_mode(true),
     );
-    let (account, account_seed) = compute_valid_account_id(account, &genesis_block_header);
+    let (account, account_seed) = compute_valid_account_id(account);
 
     let result = create_account_test(&mock_chain, account, account_seed);
 
@@ -599,15 +554,13 @@ pub fn create_account_non_fungible_faucet_invalid_initial_reserved_slot() -> any
     let mut mock_chain = MockChain::new();
     mock_chain.prove_next_block();
 
-    let genesis_block_header = mock_chain.block_header(BlockNumber::GENESIS.as_usize());
-
     let account = Account::mock_non_fungible_faucet(
         ACCOUNT_ID_PUBLIC_NON_FUNGIBLE_FAUCET,
         ZERO,
         false,
         TransactionKernel::assembler().with_debug_mode(true),
     );
-    let (account, account_seed) = compute_valid_account_id(account, &genesis_block_header);
+    let (account, account_seed) = compute_valid_account_id(account);
 
     let result = create_account_test(&mock_chain, account, account_seed);
 
@@ -625,10 +578,7 @@ pub fn create_account_invalid_seed() {
     let mut mock_chain = MockChain::new();
     mock_chain.prove_next_block();
 
-    let genesis_block_header = mock_chain.block_header(BlockNumber::GENESIS.as_usize());
-
     let (account, seed) = AccountBuilder::new(ChaCha20Rng::from_os_rng().random())
-        .anchor(AccountIdAnchor::try_from(&genesis_block_header).unwrap())
         .account_type(AccountType::RegularAccountUpdatableCode)
         .with_component(BasicWallet)
         .build()
@@ -657,7 +607,7 @@ pub fn create_account_invalid_seed() {
 
     let result = tx_context.execute_code(code);
 
-    assert_execution_error!(result, ERR_ACCOUNT_SEED_ANCHOR_BLOCK_COMMITMENT_DIGEST_MISMATCH)
+    assert_execution_error!(result, ERR_ACCOUNT_SEED_AND_COMMITMENT_DIGEST_MISMATCH)
 }
 
 #[test]
