@@ -1,23 +1,15 @@
-use alloc::{
-    boxed::Box,
-    collections::{BTreeMap, BTreeSet},
-    rc::Rc,
-    string::ToString,
-    sync::Arc,
-};
+use alloc::{boxed::Box, collections::BTreeSet, rc::Rc, sync::Arc};
 
-use miden_lib::{
-    errors::tx_kernel_errors::TX_KERNEL_ERRORS,
-    transaction::{TransactionEvent, TransactionEventError},
-};
+use miden_lib::transaction::{TransactionEvent, TransactionEventError};
 use miden_objects::{
     Digest,
     account::{AccountHeader, AccountVaultDelta},
+    assembly::mast::MastNodeExt,
 };
 use miden_tx::{TransactionMastStore, host::AccountProcedureIndexMap};
 use vm_processor::{
-    AdviceInputs, AdviceProvider, AdviceSource, ContextId, ExecutionError, Host, MastForest,
-    MastForestStore, MemAdviceProvider, ProcessState,
+    AdviceInputs, AdviceProvider, AdviceSource, ContextId, ErrorContext, ExecutionError, Host,
+    MastForest, MastForestStore, MemAdviceProvider, ProcessState,
 };
 
 // MOCK HOST
@@ -31,10 +23,6 @@ pub struct MockHost {
     adv_provider: MemAdviceProvider,
     acct_procedure_index_map: AccountProcedureIndexMap,
     mast_store: Rc<TransactionMastStore>,
-    /// Contains mappings from error codes to the related error messages.
-    ///
-    /// This map is initialized at construction time from the [`TX_KERNEL_ERRORS`] array.
-    error_messages: BTreeMap<u32, &'static str>,
 }
 
 impl MockHost {
@@ -49,13 +37,10 @@ impl MockHost {
         let adv_provider: MemAdviceProvider = advice_inputs.into();
         let proc_index_map = AccountProcedureIndexMap::new(foreign_code_commitments, &adv_provider);
 
-        let kernel_assertion_errors = BTreeMap::from(TX_KERNEL_ERRORS);
-
         Self {
             adv_provider,
             acct_procedure_index_map: proc_index_map.unwrap(),
             mast_store,
-            error_messages: kernel_assertion_errors,
         }
     }
 
@@ -70,12 +55,13 @@ impl MockHost {
     fn on_push_account_procedure_index(
         &mut self,
         process: ProcessState,
+        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
     ) -> Result<(), ExecutionError> {
         let proc_idx = self
             .acct_procedure_index_map
             .get_proc_index(&process)
-            .map_err(|err| ExecutionError::EventError(Box::new(err)))?;
-        self.adv_provider.push_stack(AdviceSource::Value(proc_idx.into()))?;
+            .map_err(|err| ExecutionError::event_error(Box::new(err), err_ctx))?;
+        self.adv_provider.push_stack(AdviceSource::Value(proc_idx.into()), err_ctx)?;
         Ok(())
     }
 }
@@ -95,37 +81,29 @@ impl Host for MockHost {
         self.mast_store.get(node_digest)
     }
 
-    fn on_event(&mut self, process: ProcessState, event_id: u32) -> Result<(), ExecutionError> {
+    fn on_event(
+        &mut self,
+        process: ProcessState,
+        event_id: u32,
+        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
+    ) -> Result<(), ExecutionError> {
         let event = TransactionEvent::try_from(event_id)
-            .map_err(|err| ExecutionError::EventError(Box::new(err)))?;
+            .map_err(|err| ExecutionError::event_error(Box::new(err), err_ctx))?;
 
         if process.ctx() != ContextId::root() {
-            return Err(ExecutionError::EventError(Box::new(
-                TransactionEventError::NotRootContext(event_id),
-            )));
+            return Err(ExecutionError::event_error(
+                Box::new(TransactionEventError::NotRootContext(event_id)),
+                err_ctx,
+            ));
         }
 
         match event {
             TransactionEvent::AccountPushProcedureIndex => {
-                self.on_push_account_procedure_index(process)
+                self.on_push_account_procedure_index(process, err_ctx)
             },
             _ => Ok(()),
         }?;
 
         Ok(())
-    }
-
-    fn on_assert_failed(&mut self, process: ProcessState, err_code: u32) -> ExecutionError {
-        let err_msg = self
-            .error_messages
-            .get(&err_code)
-            .map_or("Unknown error".to_string(), |msg| msg.to_string());
-        // Add hex representation to message so it can be easily found in MASM code.
-        let err_msg = format!("0x{:08X}: {}", err_code, err_msg);
-        ExecutionError::FailedAssertion {
-            clk: process.clk(),
-            err_code,
-            err_msg: Some(err_msg),
-        }
     }
 }
