@@ -6,11 +6,7 @@ use rand_chacha::{ChaCha20Rng, rand_core::SeedableRng};
 #[test]
 fn wallet_creation() {
     use miden_lib::account::{auth::RpoFalcon512, wallets::BasicWallet};
-    use miden_objects::{
-        account::{AccountCode, AccountStorageMode, AccountType},
-        block::BlockHeader,
-        digest,
-    };
+    use miden_objects::account::{AccountCode, AccountStorageMode, AccountType};
 
     // we need a Falcon Public Key to create the wallet account
     let seed = [0_u8; 32];
@@ -29,17 +25,8 @@ fn wallet_creation() {
     let account_type = AccountType::RegularAccountImmutableCode;
     let storage_mode = AccountStorageMode::Private;
 
-    let anchor_block_header_mock =
-        BlockHeader::mock(0, Some(digest!("0xaa")), Some(digest!("0xbb")), &[], digest!("0xcc"));
-
-    let (wallet, _) = create_basic_wallet(
-        init_seed,
-        (&anchor_block_header_mock).try_into().unwrap(),
-        auth_scheme,
-        account_type,
-        storage_mode,
-    )
-    .unwrap();
+    let (wallet, _) =
+        create_basic_wallet(init_seed, auth_scheme, account_type, storage_mode).unwrap();
 
     let expected_code = AccountCode::from_components(
         &[RpoFalcon512::new(pub_key).into(), BasicWallet.into()],
