@@ -1,29 +1,26 @@
 use alloc::{
     boxed::Box,
     collections::{BTreeMap, BTreeSet},
-    string::ToString,
     sync::Arc,
     vec::Vec,
 };
 
-use miden_lib::{
-    errors::tx_kernel_errors::TX_KERNEL_ERRORS,
-    transaction::{
-        TransactionEvent, TransactionEventError, TransactionKernelError,
-        memory::{CURRENT_INPUT_NOTE_PTR, NATIVE_NUM_ACCT_STORAGE_SLOTS_PTR},
-    },
+use miden_lib::transaction::{
+    TransactionEvent, TransactionEventError, TransactionKernelError,
+    memory::{CURRENT_INPUT_NOTE_PTR, NATIVE_NUM_ACCT_STORAGE_SLOTS_PTR},
 };
 use miden_objects::{
     Digest, Hasher,
     account::{AccountDelta, AccountHeader},
+    assembly::mast::MastNodeExt,
     asset::Asset,
     note::NoteId,
     transaction::{OutputNote, TransactionMeasurements},
     vm::RowIndex,
 };
 use vm_processor::{
-    AdviceProvider, AdviceSource, ContextId, ExecutionError, Felt, Host, MastForest,
-    MastForestStore, ProcessState,
+    AdviceProvider, AdviceSource, ContextId, ErrorContext, ExecutionError, Felt, Host, MastForest,
+    MastForestStore, MemoryError, ProcessState,
 };
 
 mod account_delta_tracker;
@@ -61,8 +58,8 @@ pub struct TransactionHost<A> {
     /// This field is updated by the [TransactionHost::on_event()] handler.
     account_delta: AccountDeltaTracker,
 
-    /// A map of the account's procedure MAST roots to the corresponding procedure indexes in the
-    /// account code.
+    /// A map of the procedure MAST roots to the corresponding procedure indices for all the
+    /// account codes involved in the transaction (for native and foreign accounts alike).
     acct_procedure_index_map: AccountProcedureIndexMap,
 
     /// The list of notes created while executing a transaction stored as note_ptr |-> note_builder
@@ -84,11 +81,6 @@ pub struct TransactionHost<A> {
     ///
     /// This field is updated by the [TransactionHost::on_trace()] handler.
     tx_progress: TransactionProgress,
-
-    /// Contains mappings from error codes to the related error messages.
-    ///
-    /// This map is initialized at construction time from the [`TX_KERNEL_ERRORS`] array.
-    error_messages: BTreeMap<u32, &'static str>,
 }
 
 impl<A: AdviceProvider> TransactionHost<A> {
@@ -98,16 +90,15 @@ impl<A: AdviceProvider> TransactionHost<A> {
         adv_provider: A,
         mast_store: Arc<dyn MastForestStore>,
         authenticator: Option<Arc<dyn TransactionAuthenticator>>,
-        mut account_code_commitments: BTreeSet<Digest>,
+        mut foreign_account_code_commitments: BTreeSet<Digest>,
     ) -> Result<Self, TransactionHostError> {
         // currently, the executor/prover do not keep track of the code commitment of the native
         // account, so we add it to the set here
-        account_code_commitments.insert(account.code_commitment());
+        foreign_account_code_commitments.insert(account.code_commitment());
 
         let proc_index_map =
-            AccountProcedureIndexMap::new(account_code_commitments, &adv_provider)?;
+            AccountProcedureIndexMap::new(foreign_account_code_commitments, &adv_provider)?;
 
-        let kernel_assertion_errors = BTreeMap::from(TX_KERNEL_ERRORS);
         Ok(Self {
             adv_provider,
             mast_store,
@@ -117,7 +108,6 @@ impl<A: AdviceProvider> TransactionHost<A> {
             authenticator,
             tx_progress: TransactionProgress::default(),
             generated_signatures: BTreeMap::new(),
-            error_messages: kernel_assertion_errors,
         })
     }
 
@@ -210,10 +200,11 @@ impl<A: AdviceProvider> TransactionHost<A> {
     fn on_account_push_procedure_index(
         &mut self,
         process: ProcessState,
+        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
     ) -> Result<(), TransactionKernelError> {
         let proc_idx = self.acct_procedure_index_map.get_proc_index(&process)?;
         self.adv_provider
-            .push_stack(AdviceSource::Value(proc_idx.into()))
+            .push_stack(AdviceSource::Value(proc_idx.into()), err_ctx)
             .expect("failed to push value onto advice stack");
         Ok(())
     }
@@ -384,6 +375,7 @@ impl<A: AdviceProvider> TransactionHost<A> {
     pub fn on_signature_requested(
         &mut self,
         process: ProcessState,
+        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
     ) -> Result<(), TransactionKernelError> {
         let pub_key = process.get_stack_word(0);
         let msg = process.get_stack_word(1);
@@ -416,7 +408,7 @@ impl<A: AdviceProvider> TransactionHost<A> {
 
         for r in signature {
             self.adv_provider
-                .push_stack(AdviceSource::Value(r))
+                .push_stack(AdviceSource::Value(r), err_ctx)
                 .map_err(|_| TransactionKernelError::FailedToPushAdviceStack(r))?;
         }
 
@@ -432,7 +424,10 @@ impl<A: AdviceProvider> TransactionHost<A> {
     /// # Errors
     /// Returns an error if the address of the currently executing input note is invalid (e.g.,
     /// greater than `u32::MAX`).
-    fn get_current_note_id(process: ProcessState) -> Result<Option<NoteId>, ExecutionError> {
+    fn get_current_note_id(
+        process: ProcessState,
+        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
+    ) -> Result<Option<NoteId>, ExecutionError> {
         // get the note address in `Felt` or return `None` if the address hasn't been accessed
         // previously.
         let note_address_felt = match process.get_mem_value(process.ctx(), CURRENT_INPUT_NOTE_PTR) {
@@ -440,9 +435,12 @@ impl<A: AdviceProvider> TransactionHost<A> {
             None => return Ok(None),
         };
         // convert note address into u32
-        let note_address: u32 = note_address_felt
-            .try_into()
-            .map_err(|_| ExecutionError::MemoryAddressOutOfBounds(note_address_felt.as_int()))?;
+        let note_address: u32 = note_address_felt.try_into().map_err(|_| {
+            ExecutionError::MemoryError(MemoryError::address_out_of_bounds(
+                note_address_felt.as_int(),
+                err_ctx,
+            ))
+        })?;
         // if `note_address` == 0 note execution has ended and there is no valid note address
         if note_address == 0 {
             Ok(None)
@@ -485,17 +483,23 @@ impl<A: AdviceProvider> Host for TransactionHost<A> {
         self.mast_store.get(node_digest)
     }
 
-    fn on_event(&mut self, process: ProcessState, event_id: u32) -> Result<(), ExecutionError> {
+    fn on_event(
+        &mut self,
+        process: ProcessState,
+        event_id: u32,
+        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
+    ) -> Result<(), ExecutionError> {
         let transaction_event = TransactionEvent::try_from(event_id)
-            .map_err(|err| ExecutionError::EventError(Box::new(err)))?;
+            .map_err(|err| ExecutionError::event_error(Box::new(err), err_ctx))?;
 
         // only the `FalconSigToStack` event can be executed outside the root context
         if process.ctx() != ContextId::root()
             && !matches!(transaction_event, TransactionEvent::FalconSigToStack)
         {
-            return Err(ExecutionError::EventError(Box::new(
-                TransactionEventError::NotRootContext(event_id),
-            )));
+            return Err(ExecutionError::event_error(
+                Box::new(TransactionEventError::NotRootContext(event_id)),
+                err_ctx,
+            ));
         }
 
         match transaction_event {
@@ -525,7 +529,7 @@ impl<A: AdviceProvider> Host for TransactionHost<A> {
             TransactionEvent::AccountAfterIncrementNonce => Ok(()),
 
             TransactionEvent::AccountPushProcedureIndex => {
-                self.on_account_push_procedure_index(process)
+                self.on_account_push_procedure_index(process,err_ctx)
             },
 
             TransactionEvent::NoteBeforeCreated => Ok(()),
@@ -534,7 +538,7 @@ impl<A: AdviceProvider> Host for TransactionHost<A> {
             TransactionEvent::NoteBeforeAddAsset => self.on_note_before_add_asset(process),
             TransactionEvent::NoteAfterAddAsset => Ok(()),
 
-            TransactionEvent::FalconSigToStack => self.on_signature_requested(process),
+            TransactionEvent::FalconSigToStack => self.on_signature_requested(process,err_ctx),
 
             TransactionEvent::PrologueStart => {
                 self.tx_progress.start_prologue(process.clk());
@@ -555,7 +559,7 @@ impl<A: AdviceProvider> Host for TransactionHost<A> {
             },
 
             TransactionEvent::NoteExecutionStart => {
-                let note_id = Self::get_current_note_id(process)?.expect(
+                let note_id = Self::get_current_note_id(process,err_ctx)?.expect(
                     "Note execution interval measurement is incorrect: check the placement of the start and the end of the interval",
                 );
                 self.tx_progress.start_note_execution(process.clk(), note_id);
@@ -584,20 +588,8 @@ impl<A: AdviceProvider> Host for TransactionHost<A> {
                 Ok(())
             }
         }
-        .map_err(|err| ExecutionError::EventError(Box::new(err)))?;
+        .map_err(|err| ExecutionError::event_error(Box::new(err),err_ctx))?;
 
         Ok(())
-    }
-
-    fn on_assert_failed(&mut self, process: ProcessState, err_code: u32) -> ExecutionError {
-        let err_msg = self
-            .error_messages
-            .get(&err_code)
-            .map_or("Unknown error".to_string(), |msg| msg.to_string());
-        ExecutionError::FailedAssertion {
-            clk: process.clk(),
-            err_code,
-            err_msg: Some(err_msg),
-        }
     }
 }
